+++ frontend/src/main.ts
@@ -13,6 +13,7 @@ import { createPinia } from 'pinia'
 import App from './App.vue'
 import router from './router'
 import { initRuntimeConfiguration, runtimeConfiguration } from '@/plugins/runtimeConfiguration'
+import { brandingPlugin } from '@/plugins/branding'
 import { httpPlugin } from '@/plugins/http'
 
 const runtimeConfigurationOptions = await initRuntimeConfiguration()
@@ -21,5 +22,6 @@ const app = createApp(App)
 app.use(createPinia())
 app.use(router)
 app.use(runtimeConfiguration, runtimeConfigurationOptions)
+app.use(brandingPlugin)
 app.use(httpPlugin)
 app.mount('#app')