+++ CHANGELOG.md
@@ -11,8 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - gateway:
   - Add support for SSO authentication with OpenID Connect (#268).
   - Add support for branding of UI with custom colors, logos and favicon (#287).
-- front: Add OIDC login page and callback view for OpenID Connect SSO
-  authentication.
+- front:
+  - Add OIDC login page and callback view for OpenID Connect SSO authentication.
+  - Support custom branding parameters as defined in UI runtime configuration
+    file.
 - Add support for Slurm REST API v0.0.45 provided by Slurm 26.05+ (#718)
 - agent: Implement (dummy) Slurm REST API adapter from v0.0.44 to v0.0.45.
 - gensession: Add `slurm-web gen-session-key` command to generate gateway